Lou_Yiping/requirements.txt
python-chess>=1.8
